Remove `log_id_cache`
The caller is responsible for ensuring that the `IdString`` parameter
exists unchanged for at least as long as the returned pointer is used.

I manually audited all (~1900) existing calls to `log_id()` and they all
satisfy this requirement. The calls fall into the following categories:
-- Parameters to log(), log_error() etc.
-- Parameters to stringf()
-- Parameters to infof()
-- Parameters to btorf_push()/btorf_pop(), where it is converted to
a std::string() immediately
-- Immediately assigning to a std::string
-- Immediately assigning to a temporary std::string to pass to a const std::string&
-- Parameters to std::string operator+, operator+=
-- Parameters to std::ostream operator<<
-- Parameters to strdup
-- Parameters to json.entry() where it is immediately converted to a std::string

Given that all existing callers satisfy the lifetime constraint even
though it wasn't required, I don't think it will be a maintenance
burden to impose this requirement going forward.

Resolves #5210.